+++ src/lib/route-data/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './routes';
 export * from './route-path';
+export * from './provider-factory';
